requirements.txt
openai>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
openenv-core>=0.2.0
uvicorn>=0.30.0